+++ winsocksample/Communication/util.h
@@ -7,7 +7,7 @@ namespace Communication {
 	
 	SOCKET createServerSocket(u_short portNum);
 	SOCKET createAcceptSocket(SOCKET listenSoc);
-	int recv_allData(SOCKET soc, char* pBuf, int32_t bufSize);
+	int recv_allData(SOCKET soc, char* pBuf, int64_t bufSize);
 
 	std::string createDispTimestamp(std::time_t* pTm);
 };
